Ignore babelrc to prevent errors when installed
When installing in another project, babel will find this babelrc and
might choose to use it for all files in the components package. This
confuses the project's babel if any plugin, ex. stage-1 is not used and
instead stage-0 is used. This should allow the consuming project to use
whatever configuration they want when this package is installed.