Update AnalyticsIcon and fix execution order
I've updated AnalyticsIcon to follow the same structure as AddMediaIcon.
I found that trying to import the PropTypes from Icon was causing the
interpreter to try and use an object that hadn't been defined. I fixed
this by removing the two lines in Icon/index.jsx that exported both
AddMediaIcon and AnalyticsIcon. This commit includes an update to the
Icon story, too.