Replace curiousBlue and denim colors
This replaces our use of 'curious-blue' with 'lochmara'. The reason for
changing this color is so that it meets the recommended contrast color
ratio of W3C. We were getting 3.45:1 with curious-blue. This change
bumps us up to 4.5:1.

Our 'denim' color is replaced by 'lochmara-hover'. Denim only appears as
a 'hover state' color, which would replace curious-blue when hovered.
This hex value was devised by adding a 10% block of black over a block
with lochmara as its background color, and then taking the result as the
value of lochmara-light. Using a color name generator, the hex value is
close to lochmara, which is why we're using the '-hover' suffix.